requirements.txt
discord.py
httpx
protobuf
python-dotenv